backend/studycountdown/views.py
from datetime import date, timedelta
from django.conf import settings as django_settings
from django.utils import timezone
from rest_framework.response import Response
from rest_framework.decorators import api_view, permission_classes, throttle_classes
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.throttling import AnonRateThrottle
from rest_framework import status
from rest_framework_simplejwt.tokens import RefreshToken

from .models import SessionNote, User, PasswordResetRequest, StudySession, UserProfile
from .serializers import (
    UserSerializer, ForgotPasswordSerializer, ResetConfirmSerializer,
    StudySessionSerializer, UserProfileSerializer
)

RESET_TOKEN_EXPIRY_MINUTES = 90

from datetime import date, timedelta
from django.conf import settings as django_settings
from django.utils import timezone
from rest_framework.response import Response
from rest_framework.decorators import api_view, permission_classes, throttle_classes
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.throttling import AnonRateThrottle
from rest_framework import status
from rest_framework_simplejwt.tokens import RefreshToken

from .models import SessionNote, User, PasswordResetRequest, StudySession, UserProfile
from .serializers import (
    UserSerializer, ForgotPasswordSerializer, ResetConfirmSerializer,
    StudySessionSerializer, UserProfileSerializer
)

RESET_TOKEN_EXPIRY_MINUTES = 90


class LoginThrottle(AnonRateThrottle):
    rate = '500/hour'


class PasswordResetThrottle(AnonRateThrottle):
    rate = '500/hour'


# Sets HTTP-only auth cookies on the response
def _set_auth_cookies(response, access_token, refresh_token=None):
    # Writes session_token (1h) and optionally refresh_token (7d) as HTTP-only cookies
    pass


# Clears all auth cookies from the response
def _delete_auth_cookies(response):
    # Deletes session_token, refresh_token, and identity cookies
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def account_registration(request):
    # Validates user data via UserSerializer
    # Creates user and sends welcome email
    # Returns 201 on success
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
@throttle_classes([LoginThrottle])
def account_login(request):
    # Validates credentials against the database
    # Issues JWT tokens as HTTP-only cookies
    # Returns is_new_user flag so the frontend can redirect to profile setup
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def token_refresh(request):
    # Reads refresh token from cookie
    # Issues new access token and rotated refresh token
    # Returns 401 and clears cookies if refresh token is invalid or expired
    pass


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def account_logout(request):
    # Blacklists the refresh token
    # Clears all auth cookies
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
@throttle_classes([PasswordResetThrottle])
def password_forget(request):
    # Validates email via ForgotPasswordSerializer
    # Generates a reset token and stores it with an expiry timestamp
    # Sends reset link via email
    # Always returns 200 to avoid revealing whether the email exists
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def password_reset(request):
    # Validates new password and token via ResetConfirmSerializer
    # Checks token exists and has not passed its expires_at time
    # Updates user password and deletes the reset token
    # Sends confirmation email
    pass


@api_view(['GET', 'POST', 'PUT'])
@permission_classes([IsAuthenticated])
def user_profile(request):
    # GET — returns current profile or 404 if not yet configured
    # POST/PUT — creates or updates profile, marks setup_complete on first save
    pass


@api_view(['GET', 'POST'])
@permission_classes([IsAuthenticated])
def study_sessions(request):
    # GET — returns available study subject choices
    # POST — saves a study session, creates a SessionNote entry if notes provided
    pass


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def session_history(request):
    # Returns all study sessions for the authenticated user, ordered by date descending
    # Prefetches related session notes
    pass


@api_view(['GET', 'DELETE'])
@permission_classes([IsAuthenticated])
def session_detail(request, session_id):
    # GET — returns a single study session
    # DELETE — deletes the session, scoped to the authenticated user
    pass


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def study_statistics(request):
    # Aggregates total hours and session count per subject for the authenticated user
    pass


# Sets HTTP-only auth cookies on the response
def _set_auth_cookies(response, access_token, refresh_token=None):
    # Writes session_token (1h) and optionally refresh_token (7d) as HTTP-only cookies
    pass


# Clears all auth cookies from the response
def _delete_auth_cookies(response):
    # Deletes session_token, refresh_token, and identity cookies
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def account_registration(request):
    # Validates user data via UserSerializer
    # Creates user and sends welcome email
    # Returns 201 on success
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
@throttle_classes([LoginThrottle])
def account_login(request):
    # Validates credentials against the database
    # Issues JWT tokens as HTTP-only cookies
    # Returns is_new_user flag so the frontend can redirect to profile setup
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def token_refresh(request):
    # Reads refresh token from cookie
    # Issues new access token and rotated refresh token
    # Returns 401 and clears cookies if refresh token is invalid or expired
    pass


@api_view(['POST'])
@permission_classes([IsAuthenticated])
def account_logout(request):
    # Blacklists the refresh token
    # Clears all auth cookies
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
@throttle_classes([PasswordResetThrottle])
def password_forget(request):
    # Validates email via ForgotPasswordSerializer
    # Generates a reset token and stores it with an expiry timestamp
    # Sends reset link via email
    # Always returns 200 to avoid revealing whether the email exists
    pass


@api_view(['POST'])
@permission_classes([AllowAny])
def password_reset(request):
    # Validates new password and token via ResetConfirmSerializer
    # Checks token exists and has not passed its expires_at time
    # Updates user password and deletes the reset token
    # Sends confirmation email
    pass


@api_view(['GET', 'POST', 'PUT'])
@permission_classes([IsAuthenticated])
def user_profile(request):
    # GET — returns current profile or 404 if not yet configured
    # POST/PUT — creates or updates profile, marks setup_complete on first save
    pass


@api_view(['GET', 'POST'])
@permission_classes([IsAuthenticated])
def study_sessions(request):
    # GET — returns available study subject choices
    # POST — saves a study session, creates a SessionNote entry if notes provided
    pass


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def session_history(request):
    # Returns all study sessions for the authenticated user, ordered by date descending
    # Prefetches related session notes
    pass


@api_view(['GET', 'DELETE'])
@permission_classes([IsAuthenticated])
def session_detail(request, session_id):
    # GET — returns a single study session
    # DELETE — deletes the session, scoped to the authenticated user
    pass


@api_view(['GET'])
@permission_classes([IsAuthenticated])
def study_statistics(request):
    # Aggregates total hours and session count per subject for the authenticated user
    pass